Refactor CreateRefPage to CreateRefFormPage and update routing. Remove unused styles and components for cleaner code. Introduce links for navigation to Create Ref Form and Drafts.
