+++ src/components/CreateRefFormPage/CreateRefFormPage.module.scss
@@ -1,0 +1,51 @@
+@import '@src/index.scss';
+
+.root {
+  display: flex;
+  align-items: stretch;
+  width: 100%;
+  margin: 0 auto;
+  height: 100%;
+  padding-top: var(--header-height);
+}
+
+.sidebar {
+  flex-basis: 160px;
+  padding: 12px;
+}
+
+.content {
+  padding: 12px;
+  overflow: auto;
+  height: 100%;
+  width: 100%;
+}
+
+.section {
+  width: 100%;
+  padding: 12px;
+}
+
+.title {
+  @include heading-200;
+  width: 100%;
+  text-align: start;
+  background-color: var(--vert-500);
+  color: white;
+  border-radius: 6px;
+  padding: 9px;
+}
+
+.field {
+  margin-bottom: 21px;
+}
+
+.row {
+  display: flex;
+  justify-content: space-between;
+  gap: 24px;
+
+  & > * {
+    flex: 1 1 0;
+  }
+}
