+++ src/components/CreateRefFormPage/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateRefFormPage';
