+++ src/components/CreateRefPage/CreateRefPage.module.scss
@@ -20,32 +20,3 @@
   height: 100%;
   width: 100%;
 }
-
-.section {
-  width: 100%;
-  padding: 12px;
-}
-
-.title {
-  @include heading-200;
-  width: 100%;
-  text-align: start;
-  background-color: var(--vert-500);
-  color: white;
-  border-radius: 6px;
-  padding: 9px;
-}
-
-.field {
-  margin-bottom: 21px;
-}
-
-.row {
-  display: flex;
-  justify-content: space-between;
-  gap: 24px;
-
-  & > * {
-    flex: 1 1 0;
-  }
-}
